+++ src/main/java/org/sopt/carena/healthreport/domain/score/metric/ScorePolicy.java
@@ -1,7 +1,6 @@
-package org.sopt.carena.healthreport.domain.score.caculator.metric;
+package org.sopt.carena.healthreport.domain.score.metric;
 
 //계산 전략 인터페이스
-
 import org.sopt.carena.healthreport.domain.score.ScoreResult;
 import org.sopt.carena.member.domain.Gender;
